+++ test/index.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const assert = require('assert')
+const { throws, strictEqual } = require('assert')
 const { spawnSync } = require('child_process')
 const path = require('path')
 const simulatePublishedTraces = require('./simulatePublishedTraces')
@@ -15,7 +15,7 @@ const tests = [
 
     const tests = new TestDirector()
     tests.add('a', () => {})
-    assert.throws(
+    throws(
       () => {
         tests.add('a', () => {})
       },
@@ -27,7 +27,7 @@ const tests = [
     console.info('Test: Test with an invalid test function type.')
 
     const tests = new TestDirector()
-    assert.throws(
+    throws(
       () => {
         tests.add('a', '')
       },
@@ -44,12 +44,12 @@ const tests = [
 
     if (error) throw error
 
-    assert.strictEqual(
+    strictEqual(
       stdout.toString(),
       '\nTest: \u001b[1ma\u001b[22m\n\n\u001b[32m\u001b[1m1/1 tests passed.\u001b[39m\u001b[22m\n\n'
     )
-    assert.strictEqual(stderr.toString(), '')
-    assert.strictEqual(status, 0)
+    strictEqual(stderr.toString(), '')
+    strictEqual(status, 0)
   },
 
   () => {
@@ -61,12 +61,12 @@ const tests = [
 
     if (error) throw error
 
-    assert.strictEqual(
+    strictEqual(
       stdout.toString(),
       '\nTest: \u001b[1ma\u001b[22m\n  Message.\n\n\u001b[32m\u001b[1m1/1 tests passed.\u001b[39m\u001b[22m\n\n'
     )
-    assert.strictEqual(stderr.toString(), '')
-    assert.strictEqual(status, 0)
+    strictEqual(stderr.toString(), '')
+    strictEqual(status, 0)
   },
 
   () => {
@@ -78,12 +78,12 @@ const tests = [
 
     if (error) throw error
 
-    assert.strictEqual(
+    strictEqual(
       stdout.toString(),
       '\nTest: \u001b[1ma\u001b[22m\n  Message A.\n\nTest: \u001b[1mb\u001b[22m\n  Message B.\n\n\u001b[32m\u001b[1m2/2 tests passed.\u001b[39m\u001b[22m\n\n'
     )
-    assert.strictEqual(stderr.toString(), '')
-    assert.strictEqual(status, 0)
+    strictEqual(stderr.toString(), '')
+    strictEqual(status, 0)
   },
 
   () => {
@@ -95,17 +95,17 @@ const tests = [
 
     if (error) throw error
 
-    assert.strictEqual(
+    strictEqual(
       stdout.toString(),
       '\nTest: \u001b[1ma\u001b[22m\n\nTest: \u001b[1mb\u001b[22m\n\nTest: \u001b[1mc\u001b[22m\n\nTest: \u001b[1md\u001b[22m\n\nTest: \u001b[1me\u001b[22m\n\nTest: \u001b[1mf\u001b[22m\n\nTest: \u001b[1mg\u001b[22m\n\nTest: \u001b[1mh\u001b[22m\n\nTest: \u001b[1mi\u001b[22m\n\n\u001b[31m\u001b[1m0/9 tests passed.\u001b[39m\u001b[22m\n\n'
     )
-    assert.strictEqual(
+    strictEqual(
       simulatePublishedTraces(stderr.toString()),
       NODE_VERSION_MAJOR < 12
-        ? "  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31m\u001b[2mtests.add (test/fixtures/fails.js:8:9)\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31mInput A expected to strictly equal input B:\n  + expected - actual\n  \n  - 0\n  + 1\u001b[39m\n  \n  \u001b[31m\u001b[2mtests.add (test/fixtures/fails.js:24:10)\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31m\u001b[2mtests.add (test/fixtures/fails.js:27:10)\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31m{ message: 'Message.' }\u001b[39m\n  \n  \u001b[31m{}\u001b[39m\n  \n  \u001b[31m'Message.'\u001b[39m\n  \n  \u001b[31mnull\u001b[39m\n"
-        : "  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31m\u001b[2mtest/fixtures/fails.js:8:9\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31mExpected values to be strictly equal:\n  \n  0 !== 1\u001b[39m\n  \n  \u001b[31m\u001b[2mtest/fixtures/fails.js:24:10\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31m\u001b[2mtest/fixtures/fails.js:27:10\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31m{ message: 'Message.' }\u001b[39m\n  \n  \u001b[31m{}\u001b[39m\n  \n  \u001b[31m'Message.'\u001b[39m\n  \n  \u001b[31mnull\u001b[39m\n"
+        ? "  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31m\u001b[2mtests.add (test/fixtures/fails.js:8:9)\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31mInput A expected to strictly equal input B:\n  + expected - actual\n  \n  - 0\n  + 1\u001b[39m\n  \n  \u001b[31m\u001b[2mtests.add (test/fixtures/fails.js:24:3)\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31m\u001b[2mtests.add (test/fixtures/fails.js:27:3)\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31m{ message: 'Message.' }\u001b[39m\n  \n  \u001b[31m{}\u001b[39m\n  \n  \u001b[31m'Message.'\u001b[39m\n  \n  \u001b[31mnull\u001b[39m\n"
+        : "  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31m\u001b[2mtest/fixtures/fails.js:8:9\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31mExpected values to be strictly equal:\n  \n  0 !== 1\u001b[39m\n  \n  \u001b[31m\u001b[2mtest/fixtures/fails.js:24:3\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31mMessage.\u001b[39m\n  \n  \u001b[31m\u001b[2mtest/fixtures/fails.js:27:3\n  Object.<anonymous> (test/fixtures/fails.js:41:7)\u001b[39m\u001b[22m\n  \n  \u001b[31m{ message: 'Message.' }\u001b[39m\n  \n  \u001b[31m{}\u001b[39m\n  \n  \u001b[31m'Message.'\u001b[39m\n  \n  \u001b[31mnull\u001b[39m\n"
     )
-    assert.strictEqual(status, 1)
+    strictEqual(status, 1)
   },
 
   () => {
@@ -117,17 +117,17 @@ const tests = [
 
     if (error) throw error
 
-    assert.strictEqual(
+    strictEqual(
       stdout.toString(),
       '\nTest: \u001b[1ma\u001b[22m\n  \n  Test: \u001b[1mb\u001b[22m\n\n\u001b[31m\u001b[1m0/1 tests passed.\u001b[39m\u001b[22m\n\n'
     )
-    assert.strictEqual(
+    strictEqual(
       simulatePublishedTraces(stderr.toString()),
       NODE_VERSION_MAJOR < 12
         ? '    \n    \u001b[31mMessage.\u001b[39m\n    \n    \u001b[31m\u001b[2mtests.add (test/fixtures/nested.js:9:11)\n    tests.add (test/fixtures/nested.js:11:15)\n    Object.<anonymous> (test/fixtures/nested.js:13:7)\u001b[39m\u001b[22m\n  \n  \u001b[31m\u001b[31m\u001b[1m0/1 tests passed.\u001b[39m\u001b[31m\u001b[22m\u001b[39m\n  \n  tests.add (test/fixtures/nested.js:11:15)\n  Object.<anonymous> (test/fixtures/nested.js:13:7)\u001b[39m\u001b[22m\n'
         : '    \n    \u001b[31mMessage.\u001b[39m\n    \n    \u001b[31m\u001b[2mtest/fixtures/nested.js:9:11\n    test/fixtures/nested.js:11:15\n    Object.<anonymous> (test/fixtures/nested.js:13:7)\u001b[39m\u001b[22m\n  \n  \u001b[31m\u001b[31m\u001b[1m0/1 tests passed.\u001b[39m\u001b[31m\u001b[22m\u001b[39m\n  \n  test/fixtures/nested.js:11:15\n  Object.<anonymous> (test/fixtures/nested.js:13:7)\u001b[39m\u001b[22m\n'
     )
-    assert.strictEqual(status, 1)
+    strictEqual(status, 1)
   },
 ]